% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/construct-areas.R
\name{construct_areas}
\alias{construct_areas}
\alias{construct_areas.default}
\alias{construct_areas.character}
\alias{construct_areas.data.frame}
\alias{construct_areas.matrix}
\title{Construct grid-template-areas property}
\usage{
construct_areas(x = NULL, ...)

\method{construct_areas}{default}(x = NULL, ...)

\method{construct_areas}{character}(x, ...)

\method{construct_areas}{data.frame}(x, ...)

\method{construct_areas}{matrix}(x, ...)
}
\arguments{
\item{x}{NULL, an atomic vector, a matrix, or a data frame}

\item{...}{discarded}
}
\description{
Construct grid-template-areas property
}
