% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-item.R
\name{grid_item}
\alias{grid_item}
\title{Construct an item to be contained by CSS grid layout}
\usage{
grid_item(..., row_start = NULL, row_end = NULL, row = NULL,
  column_start = NULL, column_end = NULL, column = NULL,
  area = NULL, justify_self = c("", "start", "center", "end"),
  align_self = c("", "start", "center", "end"), z_index = NULL,
  order = NULL, style = NULL)
}
\arguments{
\item{...}{Items in the container}

\item{row_start, row_end, row, column_start, column_end, column, area}{CSS properties of \code{gird-*}, whereas \code{grid-} is abbreviated, and hyphens are
replaced by underscores. For example, the \code{row_start} argument is equivalent
to the CSS's \code{grid-row-start} property. Specify a string as a value for the
corresponding CSS property.}

\item{justify_self, align_self}{One of \code{""}, \code{"start"}, \code{"center"}, or \code{"end"} which can be partial matched.
\code{""} does not specify the corresponding CSS property (\code{justify-self}, and
\code{align-self}, respctively).}

\item{z_index, order}{A string to specify the corresponding CSS property (\code{z-index} and \code{order},
respectively)}

\item{style}{Additional values of the style attribute such as \code{color: red;}.}
}
\description{
Construct an item to be contained by CSS grid layout
}
