% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-rowwise.R, R/grid-colwise.R
\name{grid_rowwise}
\alias{grid_rowwise}
\alias{grid_colwise}
\title{Gird layout in single row or column}
\usage{
grid_rowwise(..., cols = character(0L))

grid_colwise(..., rows = character(0L))
}
\arguments{
\item{...}{Items in CSS Grid Layout and arguments passed to \code{\link{grid_layout}}.}

\item{rows, cols}{Sizes of rows and columns in a character vector.
If the given number of sizes are less than the number of items,
then \code{"auto"} is used for items missing sizes.}
}
\description{
Gird layout in single row or column
}
