% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid-item.R
\name{grid_item}
\alias{grid_item}
\title{Construct an item to be contained by CSS grid layout}
\usage{
grid_item(..., row_start = NULL, row_end = NULL, row = NULL,
  column_start = NULL, column_end = NULL, column = NULL,
  area = NULL, justify_self = NULL, align_self = NULL,
  z_index = NULL, order = NULL, style = NULL)
}
\arguments{
\item{...}{Items in the container}

\item{row_start, row_end, row, column_start, column_end, column, area}{CSS properties of \code{gird-*}, whereas \code{grid-} is abbreviated, and hyphens are
replaced by underscores. For example, the \code{row_start} argument is equivalent
to the CSS's \code{grid-row-start} property. Specify a string as a value for the
corresponding CSS property.}

\item{justify_self, align_self}{Strings to specify values for the CSS's
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/justify-self}{justify-self}
and
\href{https://developer.mozilla.org/en-US/docs/Web/CSS/align-self}{align-self}
properties.}

\item{z_index, order}{A string to specify the corresponding CSS property (\code{z-index} and \code{order},
respectively)}

\item{style}{Additional values of the style attribute such as \code{color: red;}.}
}
\description{
Construct an item to be contained by CSS grid layout
}
